LayoutTests/http/tests/security/dataURL/xss-DENIED-to-data-url-from-data-url-expected.txt

CONSOLE MESSAGE: line 1: Unsafe JavaScript attempt to access frame with URL data:text/html,%3Chtml%3E%3Cbody%3E%3Cp%20id='accessMe'%3EPASS:%20Cross%20frame%20access%20from%20a%20data:%20URL%20was%20denied.%3C/p%3E%3Cp%3EInner%20iframe.%3C/p%3E%3Ciframe%20name='innerFrame'%20id='innerFrame'%20src='data:text/html,%3Chtml%3E%3Chead%3E%3Cscript%3Efunction%20test()%20{try%20{parent.document.getElementById(%22accessMe%22).innerHTML%20=%20%22FAIL:%20Cross%20frame%20access%20from%20a%20data:%20URL%20was%20allowed.%22;}%20catch%20(e)%20{}if%20(window.layoutTestController)layoutTestController.notifyDone();}%3C/script%3E%3C/head%3E%3Cbody%20onload=%22test();%22%3E%3Cp%3EInner-inner%20iframe.%3C/p%3E%3C/body%3E%3C/html%3E'%3E%3C/iframe%3E%3C/body%3E%3C/html%3E from frame with URL data:text/html,%3Chtml%3E%3Chead%3E%3Cscript%3Efunction%20test()%20{try%20{parent.document.getElementById(%22accessMe%22).innerHTML%20=%20%22FAIL:%20Cross%20frame%20access%20from%20a%20data:%20URL%20was%20allowed.%22;}%20catch%20(e)%20{}if%20(window.layoutTestController)layoutTestController.notifyDone();}%3C/script%3E%3C/head%3E%3Cbody%20onload=%22test();%22%3E%3Cp%3EInner-inner%20iframe.%3C/p%3E%3C/body%3E%3C/html%3E. Domains, protocols and ports must match.

This tests that a data: URL loaded in an iframe inside another data: URL loaded iframe doesn't have access to its parent, the first data: URL loaded iframe.



--------
Frame: 'aFrame'
--------
PASS: Cross frame access from a data: URL was denied.

Inner iframe.



--------
Frame: 'innerFrame'
--------
Inner-inner iframe.
